apply striped table effect, refactor code

+++ Program.cs
@@ -15,14 +15,12 @@ namespace QPharma
             var runningProcess = Process.GetProcessesByName(currentProcess.ProcessName)
                 .FirstOrDefault(p => p.Id != currentProcess.Id);
 
-            if (runningProcess != null)
-            {
-                runningProcess.Kill();
-            }
+            if (runningProcess != null) runningProcess.Kill();
 
             Application.EnableVisualStyles();
             Application.SetCompatibleTextRenderingDefault(false);
             Application.Run(new SplashGUI());
+            
         }
     }
 }
